Organized folder structure for client side components.
