+++ FribergFastigheter.Client/Components/Housing/HousingDetails.razor
