+++ FribergFastigheter.Client/Components/Housing/HousingDetails.razor.css
