+++ FribergFastigheter.Client/Components/Housing/HousingListItem.razor.css
