+++ FribergFastigheter.Client/Pages/BrokerDetailsPage.razor
@@ -1,5 +1,7 @@
 ﻿@page "/broker/{BrokerId:int}"
 @using FribergFastigheter.Client.Components
+@using FribergFastigheter.Client.Components.Broker
+@using FribergFastigheter.Client.Components.Housing
 @*
 <!-- Author: Jimmie -->
 <!-- Co Authors: -->
