+++ FribergFastigheter.Client/Pages/BrokerFirmMember/Housing/ListHousingPage.razor
@@ -1,5 +1,6 @@
 ﻿@page "/brokermember/housing/list"
 @using FribergFastigheter.Client.Components
+@using FribergFastigheter.Client.Components.Housing
 @using Microsoft.AspNetCore.Authorization
 @*
     <!-- Author: Jimmie -->
