+++ FribergFastigheter.Client/Pages/BrokerFirmPage.razor
@@ -1,5 +1,6 @@
 ﻿@page "/brokerfirm/{BrokerFirmId:int}"
 @using FribergFastigheter.Client.Components
+@using FribergFastigheter.Client.Components.Broker
 @*
 <!-- Author: Jimmie -->
 <!-- Co Authors: -->
