+++ FribergFastigheter.Client/Pages/Home.razor
@@ -1,5 +1,6 @@
 ﻿@page "/"
 @using FribergFastigheter.Client.Components
+@using FribergFastigheter.Client.Components.Housing
 @* 
 <!-- Author: Jimmie -->
 <!-- Co Authors: -->
