+++ FribergFastigheter.Client/Pages/HousingDetailsPage.razor
@@ -1,5 +1,6 @@
 ﻿@page "/housing/{id:int}"
 @using FribergFastigheter.Client.Components
+@using FribergFastigheter.Client.Components.Housing
 
 <div class="vertical-page-container">
 	@if (_housingViewModel != null)
